requirements.txt
# Django + API
Django>=4.2,<4.3
djangorestframework>=3.15,<4.0

# Celery + Redis
celery>=5.3,<6.0
redis>=5.0,<6.0

# Postgres
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.0,<3.0

# WSGI для запуска в Docker
gunicorn>=22.0,<23.0

# Тесты
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-celery>=1.0,<2.0